+++ CHANGELOG.rst
@@ -16,6 +16,7 @@ Changelog
 * Update :command:`import-read-depth` command to be able to subset/exclude specified samples.
 * Rename ``--samples`` argument from :command:`compute-copy-number` command to ``--samples-without-sv``.
 * Rename ``--samples`` argument from :command:`run-ngs-pipeline` command to ``--samples-without-sv``.
+* Update :command:`run-ngs-pipeline` and :command:`run-chip-pipeline` commands to be able to subset/exclude specified samples.
 
 0.9.0 (2021-12-05)
 ------------------